backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.2
black==23.9.1
flake8==6.1.0 